Feat: feat: UnitStatCalculator 추가 및 Stat,Training 로직 정비

+++ Assets/2.Scripts/Unit/Model/Skill/HealSkillData.cs
@@ -10,8 +10,8 @@ public class HealSkillData : SkillData
     public override void Use(UnitController caster, List<UnitController> targets)
     {
         // TODO Heal 타겟잡는것 수정 -> SelfCircleScanner에서 조금 더 변형시켜야함 -> HealEffect를 조절해야함. (public Vector2 TargetPos => targetPos; 를 List형으로 만들어서 전체힐로도 할 수 있어야할 듯)
-        int healAmount = CalculateHealAmount(caster);
-        int minHp = int.MaxValue;
+        long healAmount = CalculateHealAmount(caster);
+        long minHp = int.MaxValue;
         UnitController minHpTarget = null;
         
         foreach (UnitController target in targets)
